ERP_DATA_ANALYZER/requirements.txt
streamlit
sqlalchemy
pandas
plotly
watchdog
python-dotenv
altair